requirements.txt
numpy
pandas
scikit-learn
scipy
pyyaml
pydantic
joblib
tqdm
torch
torchvision
timm
Pillow
opencv-python-headless
scikit-image
albumentations
shap
captum
grad-cam
streamlit
uvicorn
python-multipart
httpx
Jinja2
reportlab
cairosvg
matplotlib
plotly
seaborn
pytest
ruff
